+++ db/04_pago.sql
@@ -1,7 +1,7 @@
 --Medio de Pago
 
 CREATE TABLE pago.medio_pago (
-	id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
 	tipo varchar(100) DEFAULT '' NOT NULL,
 	nombre varchar(100) DEFAULT '' NOT NULL,
 	CONSTRAINT medio_pago_pk PRIMARY KEY (id)
@@ -18,7 +18,7 @@ COMMENT ON COLUMN pago.medio_pago.nombre IS 'Nombre del medio de pago, por ejemp
 --Promoción
 
 CREATE TABLE pago.promocion (
-  id numeric DEFAULT 0 NOT NULL,
+  id integer DEFAULT 0 NOT NULL,
   porcentaje decimal DEFAULT 0 NOT NULL,
   codigo varchar(50) DEFAULT '' NOT NULL,
   esta_disponible boolean DEFAULT false NOT NULL,
@@ -35,10 +35,10 @@ COMMENT ON COLUMN pago.promocion.esta_disponible IS 'Disponible (true), No Dispo
 --Pago
 
 CREATE TABLE pago.pago (
-  id numeric DEFAULT 0 NOT NULL,
-  viaje_id numeric DEFAULT 0 NOT NULL,
-  medio_pago_id numeric DEFAULT 0 NOT NULL,
-  promocion_id numeric DEFAULT 0 NOT NULL,
+  id integer DEFAULT 0 NOT NULL,
+  viaje_id integer DEFAULT 0 NOT NULL,
+  medio_pago_id integer DEFAULT 0 NOT NULL,
+  promocion_id integer DEFAULT 0 NOT NULL,
   valor_total decimal DEFAULT 0 NOT NULL,
   comision decimal DEFAULT 0 NOT NULL,
   valor_impuestos decimal DEFAULT 0 NOT NULL,
@@ -61,8 +61,8 @@ COMMENT ON COLUMN pago.pago.valor_impuestos IS 'Define el valor del impuesto que
 --Referido
 
 CREATE TABLE pago.referido (
-  id numeric DEFAULT 0 NOT NULL,
-  usuario_id numeric DEFAULT 0 NOT NULL,
+  id integer DEFAULT 0 NOT NULL,
+  usuario_id integer DEFAULT 0 NOT NULL,
   correo varchar(250) DEFAULT '' NOT NULL,
   telefono varchar(30) DEFAULT '' NOT NULL,
   CONSTRAINT referido_pk PRIMARY KEY (id),
