+++ db/init.sql
@@ -6,7 +6,7 @@ CREATE SCHEMA soporte;
 -- Tipo de Identificación
 
 CREATE TABLE "seguridad".tipo_identificacion (
-	id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
 	nombre varchar(40) DEFAULT '' NOT NULL,
 	CONSTRAINT tipoIdentificacion_pk PRIMARY KEY (id)
 );
@@ -21,7 +21,7 @@ COMMENT ON COLUMN "seguridad".tipo_identificacion.nombre IS 'Nombre del tipo de 
 --Rol
 
 CREATE TABLE seguridad.rol (
-	id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
 	nombre varchar(100) DEFAULT '' NOT NULL,
 	CONSTRAINT rol_pk PRIMARY KEY (id)
 );
@@ -36,9 +36,9 @@ COMMENT ON COLUMN seguridad.rol.nombre IS 'Nombre del rol';
 --Usuario
 
 CREATE TABLE seguridad.usuario (
-	id numeric DEFAULT 0 NOT NULL,
-	tipo_identificacion_id numeric DEFAULT 0 NOT NULL,
-	rol_id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
+	tipo_identificacion_id integer DEFAULT 0 NOT NULL,
+	rol_id integer DEFAULT 0 NOT NULL,
 	identificacion varchar(20) DEFAULT '' NOT NULL,
 	nombre varchar(100) DEFAULT '' NOT NULL,
 	apellido varchar(100) DEFAULT '' NOT NULL,
@@ -65,9 +65,9 @@ COMMENT ON COLUMN seguridad.usuario."password" IS 'Contraseña usuario';
 
 
 CREATE TABLE transporte."tipo_vehiculo" (
-	id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
 	tipo varchar(100) DEFAULT '' NOT NULL,
-	cantidad_de_ejes numeric DEFAULT 0 NOT NULL,
+	cantidad_de_ejes integer DEFAULT 0 NOT NULL,
 	CONSTRAINT tipo_de_vehiculo_pk PRIMARY KEY (id)
 );
 COMMENT ON TABLE transporte."tipo_vehiculo" IS 'Almacena los tipos de vehículos';
@@ -83,8 +83,8 @@ COMMENT ON COLUMN transporte."tipo_vehiculo".cantidad_de_ejes IS 'define la cant
 
 
 CREATE TABLE transporte.conductor (
-	id numeric DEFAULT 0 NOT NULL,
-	usuario_id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
+	usuario_id integer DEFAULT 0 NOT NULL,
 	numero_licencia varchar(200) DEFAULT '' NOT NULL,
 	CONSTRAINT conductor_pk PRIMARY KEY (id),
 	CONSTRAINT conductor_usuario_fk FOREIGN KEY (usuario_id) REFERENCES seguridad.usuario(id)
@@ -102,13 +102,13 @@ COMMENT ON COLUMN transporte.conductor.numero_licencia IS 'numero de licencia de
 
 
 CREATE TABLE transporte.vehiculo (
-	id numeric DEFAULT 0 NOT NULL,
-	conductor_id numeric DEFAULT 0 NOT NULL,
-	tipo_vehículo_id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
+	conductor_id integer DEFAULT 0 NOT NULL,
+	tipo_vehículo_id integer DEFAULT 0 NOT NULL,
 	placa varchar(20) DEFAULT '' NOT NULL,
 	color varchar(20) DEFAULT '' NOT NULL,
 	modelo varchar(20) DEFAULT '' NOT NULL,
-	asientos numeric DEFAULT 0 NOT NULL,
+	asientos integer DEFAULT 0 NOT NULL,
 	esta_disponible boolean DEFAULT false NOT NULL,
 	CONSTRAINT vehiculo_pk PRIMARY KEY (id),
 	CONSTRAINT vehiculo_conductor_fk FOREIGN KEY (conductor_id) REFERENCES transporte.conductor(id),
@@ -131,10 +131,10 @@ COMMENT ON COLUMN transporte.vehiculo.esta_disponible IS 'disponible (1), no dis
 -- Solicitante
 
 CREATE TABLE transporte.solicitante (
-	id numeric DEFAULT 0 NOT NULL,
-	usuario_id numeric DEFAULT 0 NULL,
-	latitud_defecto numeric DEFAULT 0 NOT NULL,
-	longitud_defecto numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
+	usuario_id integer DEFAULT 0 NULL,
+	latitud_defecto integer DEFAULT 0 NOT NULL,
+	longitud_defecto integer DEFAULT 0 NOT NULL,
 	CONSTRAINT solicitante_pk PRIMARY KEY (id),
 	CONSTRAINT solicitante_usuario_fk FOREIGN KEY (usuario_id) REFERENCES seguridad.usuario(id)
 );
@@ -151,14 +151,14 @@ COMMENT ON COLUMN transporte.solicitante.longitud_defecto IS 'Define coordenada 
 --Viaje
 
 CREATE TABLE transporte.viaje (
-	id numeric DEFAULT 0 NOT NULL,
-	solicitante_id numeric DEFAULT 0 NOT NULL,
-	vehiculo_id numeric DEFAULT 0 NOT NULL,
-	partida_latitud numeric DEFAULT 0 NOT NULL,
-	partida_longitud numeric DEFAULT 0 NOT NULL,
-	destino_latitud numeric DEFAULT 0 NOT NULL,
-	destino_longitud numeric DEFAULT 0 NOT NULL,
-	calificacion numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
+	solicitante_id integer DEFAULT 0 NOT NULL,
+	vehiculo_id integer DEFAULT 0 NOT NULL,
+	partida_latitud integer DEFAULT 0 NOT NULL,
+	partida_longitud integer DEFAULT 0 NOT NULL,
+	destino_latitud integer DEFAULT 0 NOT NULL,
+	destino_longitud integer DEFAULT 0 NOT NULL,
+	calificacion integer DEFAULT 0 NOT NULL,
 	observaciones varchar(1000) DEFAULT '' NOT NULL,
 	CONSTRAINT viaje_pk PRIMARY KEY (id),
 	CONSTRAINT viaje_solicitante_fk FOREIGN KEY (solicitante_id) REFERENCES transporte.solicitante(id),
@@ -182,7 +182,7 @@ COMMENT ON COLUMN transporte.viaje.observaciones IS 'Especifica las observacione
 --Medio de Pago
 
 CREATE TABLE pago.medio_pago (
-	id numeric DEFAULT 0 NOT NULL,
+	id integer DEFAULT 0 NOT NULL,
 	tipo varchar(100) DEFAULT '' NOT NULL,
 	nombre varchar(100) DEFAULT '' NOT NULL,
 	CONSTRAINT medio_pago_pk PRIMARY KEY (id)
@@ -199,7 +199,7 @@ COMMENT ON COLUMN pago.medio_pago.nombre IS 'Nombre del medio de pago, por ejemp
 --Promoción
 
 CREATE TABLE pago.promocion (
-  id numeric DEFAULT 0 NOT NULL,
+  id integer DEFAULT 0 NOT NULL,
   porcentaje decimal DEFAULT 0 NOT NULL,
   codigo varchar(50) DEFAULT '' NOT NULL,
   esta_disponible boolean DEFAULT false NOT NULL,
@@ -216,10 +216,10 @@ COMMENT ON COLUMN pago.promocion.esta_disponible IS 'Disponible (true), No Dispo
 --Pago
 
 CREATE TABLE pago.pago (
-  id numeric DEFAULT 0 NOT NULL,
-  viaje_id numeric DEFAULT 0 NOT NULL,
-  medio_pago_id numeric DEFAULT 0 NOT NULL,
-  promocion_id numeric DEFAULT 0 NOT NULL,
+  id integer DEFAULT 0 NOT NULL,
+  viaje_id integer DEFAULT 0 NOT NULL,
+  medio_pago_id integer DEFAULT 0 NOT NULL,
+  promocion_id integer DEFAULT 0 NOT NULL,
   valor_total decimal DEFAULT 0 NOT NULL,
   comision decimal DEFAULT 0 NOT NULL,
   valor_impuestos decimal DEFAULT 0 NOT NULL,
@@ -242,8 +242,8 @@ COMMENT ON COLUMN pago.pago.valor_impuestos IS 'Define el valor del impuesto que
 --Referido
 
 CREATE TABLE pago.referido (
-  id numeric DEFAULT 0 NOT NULL,
-  usuario_id numeric DEFAULT 0 NOT NULL,
+  id integer DEFAULT 0 NOT NULL,
+  usuario_id integer DEFAULT 0 NOT NULL,
   correo varchar(250) DEFAULT '' NOT NULL,
   telefono varchar(30) DEFAULT '' NOT NULL,
   CONSTRAINT referido_pk PRIMARY KEY (id),
@@ -261,8 +261,8 @@ COMMENT ON COLUMN pago.referido.telefono IS 'Almacena el número de teléfono de
 --Agente
 
 CREATE TABLE soporte.agente (
-id numeric DEFAULT 0 NOT NULL,
-usuario_id numeric DEFAULT 0 NOT NULL,
+id integer DEFAULT 0 NOT NULL,
+usuario_id integer DEFAULT 0 NOT NULL,
 alias varchar(100) DEFAULT '' NOT NULL,
 CONSTRAINT agente_pk PRIMARY KEY (id),
 CONSTRAINT agente_usuario_fk FOREIGN KEY (usuario_id) REFERENCES
@@ -278,8 +278,8 @@ COMMENT ON COLUMN soporte.agente.alias IS 'Define el nombre que se le presenta a
 --pqrs
 
 CREATE TABLE soporte.pqrs (
-id numeric DEFAULT 0 NOT NULL,
-agente_id numeric DEFAULT 0 NOT NULL,
+id integer DEFAULT 0 NOT NULL,
+agente_id integer DEFAULT 0 NOT NULL,
 tipo varchar(1) DEFAULT '' NOT NULL,
 estado varchar(60) DEFAULT '' NOT NULL,
 observaciones varchar(1000) NOT NULL,
